Extract element stack to a separate class
